refactor: move track to a new domain on frontend
